Working on the original xbox !
- Changed the syntax of some of the code to work on old compilers
- Added the xbox version of the SDL 1.2

+++ GAME_Functions.c
@@ -4,12 +4,12 @@ void BallRespown(Vector2f* BallSpeed, float NominalSpeed){
     float Direction;
     Vector2f TempSpeed;
 
-    Direction = 60 - (rand()%121); // entre 0 et 120
+    Direction = 60.0f - (float)(rand()%121); // entre 0 et 120
     Direction = Direction*3.14f/180.0f; // Deg to Rad
     TempSpeed.y = -NominalSpeed;
     TempSpeed.x = 0;
 
     // Rotation Matrix
-    BallSpeed->x = TempSpeed.x * cos(Direction) - TempSpeed.y * sin(Direction);
-    BallSpeed->y = TempSpeed.x * sin(Direction) + TempSpeed.y * cos(Direction);
+    BallSpeed->x = TempSpeed.x * (float)cos(Direction) - TempSpeed.y * (float)sin(Direction);
+    BallSpeed->y = TempSpeed.x * (float)sin(Direction) + TempSpeed.y * (float)cos(Direction);
 }
